+++ ops-work-order-system/src/main/java/com/example/opsworkordersystem/repository/ApprovalRecordRepository.java
@@ -10,9 +10,9 @@ import java.util.List;
 @Repository
 public interface ApprovalRecordRepository extends JpaRepository<ApprovalRecord, Long> {
     // 根据工单ID查询审批记录
-    List<ApprovalRecord> findByWorkOrderId(Long workOrderId);
+    List<ApprovalRecord> findByWorkOrderId(Integer workOrderId);
     
     // 预加载审批人的查询
     @Query("SELECT a FROM ApprovalRecord a LEFT JOIN FETCH a.approver WHERE a.workOrder.id = ?1")
-    List<ApprovalRecord> findByWorkOrderIdWithApprover(Long workOrderId);
+    List<ApprovalRecord> findByWorkOrderIdWithApprover(Integer workOrderId);
 }
